refactor imports in CodexPaneSection and PortfolioSection, update primaryAction hrefs, and rename ProjectSection component
